This is an upload of my first lab
For most of activities/labs that I upload, they will all be based on the Jupiter Lab Notebook but all the code will be Python based. For all of these activities, you'll see the examples that my professor provided and if you scroll down, you'll eventually see the actual code that I had to do for the activities.