Lab 2
This lab was a bit of a heavy one. We were told to create a program that would allow the user to input a YouTube link of their choosing and output the name of the view, the description, and name of the channel. Also, we had to create a function that would take in a string of text and then output all video chapters that are in that text as a tuple of strings.