Lab 3
For this lab, we have to use the code from our previous lab (lab 2) and convert it into a class that will hold and operate on the data outputted from lab 2. We also have to create another class function that will save and load an object to/from files using json libraries.