Final Project!
For this final project, we had to create our own original app using Tkinter. The difference between this app and the other app from activity 9, we had to add more widgets, additional GUI features, and compile the app as an executable. Thankfully I was able to compile my app as an executable and I just hopes that it works on other people's end.